+++ EduConnect-FrontEnd/src/components/privado/Admin/FormularioAdmin.jsx
@@ -1,10 +1,10 @@
 import Button from 'react-bootstrap/Button'
 import Form from 'react-bootstrap/Form'
 import Card from 'react-bootstrap/Card'
-import useMostrarAlertaConfirmacion from '../../hooks/useMostrarAlertaConfirmacion'
-import '../styles/privado/formulario.css'
+import useMostrarAlertaConfirmacion from '../../../hooks/useMostrarAlertaConfirmacion'
+import '../../../styles/privado/formulario.css'
 
-function FormularioEstudiante () {
+function FormularioAdmin () {
   return (
         <>
         <Card className=" formulario">
@@ -54,4 +54,4 @@ function FormularioEstudiante () {
   )
 }
 
-export default FormularioEstudiante
+export default FormularioAdmin
