+++ EduConnect-FrontEnd/src/components/privado/SuperAdmin/AsideSuperAdmin.jsx
@@ -1,25 +1,25 @@
 import { Col, Container, Row } from 'react-bootstrap'
-import '../styles/privado/aside.css'
+import '../../../styles/privado/aside.css'
 import Dropdown from 'react-bootstrap/Dropdown'
 import DropdownButton from 'react-bootstrap/DropdownButton'
 import Image from 'react-bootstrap/Image'
 
-function AsideAdmin () {
+function AsideSuperAdmin () {
   return (
         <>
             <div className="aside">
                 <div className='perfil'>
                     <Image src="https://es.seaicons.com/wp-content/uploads/2015/06/user-male-icon.png" roundedCircle />
-                    <h3>Admin</h3>
+                    <h3>SuperAdmin</h3>
                     <p>Profile</p>
                 </div>
                 <Container fluid>
                     <Row>
                         <Col md={{ span: 6, offset: 4 }} >
-                            <DropdownButton id="dropdown-basic-button2" title="Selecciona" drop={'down-centered'} >
-                                <Dropdown.Item href='/admin/formularioProfesor'>Profesor</Dropdown.Item>
-                                <Dropdown.Item href="/admin/formularioEstudiante">Estudiante</Dropdown.Item>
-                                <Dropdown.Item href="#/admin/asignarCurso">Curso</Dropdown.Item>
+                            <DropdownButton id="dropdown-basic-button" title="Selecciona" drop={'down-centered'} >
+                                <Dropdown.Item href='/superAdmin/formularioColegio'>Colegio</Dropdown.Item>
+                                <Dropdown.Item href="/superAdmin/formularioAdmin">Administrador</Dropdown.Item>
+                                <Dropdown.Item href="#/superAdmin/asignarColegio">Asignar Colegio</Dropdown.Item>
                             </DropdownButton>
                         </Col>
                     </Row>
@@ -28,4 +28,4 @@ function AsideAdmin () {
         </>
   )
 }
-export default AsideAdmin
+export default AsideSuperAdmin
